Defining a list of shape functions for each triang
Changed my mind. I will define a list of shape function objects for each
triangle. Each shape function object will contain phi and grad phi
corresponding to the corresponding DOF in that triangle. I know this is
far from optimal, but this should still work for the integration over
the reference triangle by composing everything with the affine mapping
to evaluate everything in the Gauss-Legendre quadrature over the
reference triangle.